rename packages - part 2
